# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Authentication API",
    "version": "1.0.0",
    "description": "Learning Management System Authentication API with comprehensive user management features",
    "contact": {
      "name": "API Support",
      "email": "support@lms.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Request success status"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          }
        }
      },
...
      "FriendRequest": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "integer",
            "description": "Friend request ID"
          },
          "senderId": {
            "type": "integer",
            "description": "Sender user ID"
          },
          "receiverId": {
            "type": "integer",
            "description": "Receiver user ID"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected",
              "blocked"
            ],
            "description": "Request status"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "description": "Request sent timestamp"
          }
        }
      },
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {}
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/friends/request": {
      "post": {
        "summary": "Send a friend request to another user",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendFriendRequestRequest"
              },
              "example": {
                "receiverId": 123,
                "message": "Hi! I'd like to connect and study together."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Friend request sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or already friends/blocked"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Friend request already exists"
          }
        }
      }
    },
    "/friends/request/{id}/accept": {
      "post": {
        "summary": "Accept a friend request",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Friend request ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend request accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to accept this request"
          },
          "404": {
            "description": "Friend request not found"
          }
        }
      }
    },
    "/friends/request/{id}/reject": {
      "post": {
        "summary": "Reject a friend request",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Friend request ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend request rejected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to reject this request"
          },
          "404": {
            "description": "Friend request not found"
          }
        }
      }
    },
    "/friends": {
      "get": {
        "summary": "Get user's friends list with pagination",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of friends per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search friends by name or email"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "offline",
                "all"
              ],
              "default": "all"
            },
            "description": "Filter by online status"
          }
        ],
        "responses": {
          "200": {
            "description": "Friends list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Friend"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/friends/requests": {
      "get": {
        "summary": "Get pending friend requests (sent and received)",
        "tags": [
          "Friends"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of requests per page"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "sent",
                "received",
                "all"
              ],
              "default": "all"
            },
            "description": "Filter by request type"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "accepted",
                "rejected"
              ],
              "default": "pending"
            },
            "description": "Filter by request status"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend requests retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/FriendRequest"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}